+++ cli/Valet/Facades/Adminer.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Valet\Facades;
+
+/**
+ * Class Adminer.
+ *
+ * @method static void install()
+ * @method static void ensureInstalled()
+ * @method static void uninstall()
+ * @method static string path()
+ * @method static string pluginsPath()
+ * @method static string enabledPluginsFile()
+ * @method static string url()
+ * @method static void status(bool $open = false, bool $pathOnly = false)
+ */
+class Adminer extends Facade
+{
+}