+++ cli/stubs/adminer/plugins/enabled.php
@@ -1,0 +1,15 @@
+<?php
+
+/**
+ * Enable Adminer plugins by returning instances below.
+ * Place plugin PHP files next to this file, then add them here.
+ *
+ * @see https://www.adminer.org/en/plugins/
+ *
+ * @return list<object>
+ */
+return [
+    // new AdminerTablesFilter(),
+    // new AdminerDumpJson(),
+    // new AdminerEditTextarea(),
+];